Attempt to fix build by renaming module
